// src/app/common/models/guest-info-details.model.ts
import { Preferences } from "./preferences.model";

export class GuestInfoDetails {
  salutation: string;
  firstName: string;
  lastName: string;
  streetAddress1: string;
  streetAddress2: string;
  city: string;
  state: string;
  countryCode: string;
  phoneNumber: string;
  emailAddress: string;
  callingCode: string;
  countryName: string;
  postalCode: string;
}